--- a/spec/rspec/core/bisect/coordinator_spec.rb
+++ b/spec/rspec/core/bisect/coordinator_spec.rb
@@ -6,7 +6,9 @@ module RSpec::Core
   RSpec.describe Bisect::Coordinator, :simulate_shell_allowing_unquoted_ids do
     include FormatterSupport
 
-    let(:fake_runner) do
+    let(:config) { instance_double(Configuration, :bisect_runner_class => fake_bisect_runner) }
+    let(:spec_runner) { instance_double(RSpec::Core::Runner, :configuration => config) }
+    let(:fake_bisect_runner) do
       FakeBisectRunner.new(
         1.upto(8).map { |i| "#{i}.rb[1:1]" },
         %w[ 2.rb[1:1] ],
@@ -15,12 +17,7 @@ module RSpec::Core
     end
 
     def find_minimal_repro(output, formatter=Formatters::BisectProgressFormatter)
-      allow(Bisect::Server).to receive(:run).and_yield(instance_double(Bisect::Server))
-      allow(Bisect::ShellRunner).to receive(:new).and_return(fake_runner)
-
-      Bisect::Coordinator.bisect_with([], formatter.new(output))
-    ensure
-      RSpec.reset # so that RSpec.configuration.output_stream isn't closed
+      Bisect::Coordinator.bisect_with(spec_runner, [], formatter.new(output))
     end
 
     it 'notifies the bisect progress formatter of progress and closes the output' do
@@ -52,7 +49,7 @@ module RSpec::Core
       output = normalize_durations(output.string)
 
       expect(output).to eq(<<-EOS.gsub(/^\s+\|/, ''))
-        |Bisect started using options: ""
+        |Bisect started using options: "" and bisect runner: FakeBisectRunner
         |Running suite to find failures... (n.nnnn seconds)
         | - Failing examples (2):
         |    - 2.rb[1:1]
@@ -100,7 +97,7 @@ module RSpec::Core
 
     context "with an order-independent failure" do
       it "detects the independent case and prints the minimal reproduction" do
-        fake_runner.dependent_failures = {}
+        fake_bisect_runner.dependent_failures = {}
         output = StringIO.new
         find_minimal_repro(output)
         output = normalize_durations(output.string)
@@ -120,13 +117,13 @@ module RSpec::Core
       end
 
       it "can use the debug formatter for detailed output" do
-        fake_runner.dependent_failures = {}
+        fake_bisect_runner.dependent_failures = {}
         output = StringIO.new
         find_minimal_repro(output, Formatters::BisectDebugFormatter)
         output = normalize_durations(output.string)
 
         expect(output).to eq(<<-EOS.gsub(/^\s+\|/, ''))
-          |Bisect started using options: ""
+          |Bisect started using options: "" and bisect runner: FakeBisectRunner
           |Running suite to find failures... (n.nnnn seconds)
           | - Failing examples (1):
           |    - 2.rb[1:1]