… Changelog.md
… spec/integration/bisect_runners_spec.rb
… spec/rspec/core/configuration_spec.rb
… spec/rspec/core/invocations_spec.rb
… spec/rspec/core/resources/inconsistently_ordered_specs.rb