@myronmarston 2018-01-27
This is green and ready for review, @rspec/rspec.