@JonRowe 2018-01-27
Looks good to me! My 2¢ on whether we should default to this is that we should, but if you're concerned we could do a deprecation style warning for not setting which one you're using and explain the choice?