@myronmarston 2018-01-27
> Looks good to me!

What, no suggestions whatsoever?   Is that a first for a PR of this size? 😄 

> but if you're concerned

I'm _vaguely_ concerned--but mostly it's a concern about the fact that we don't yet know how frequently the fork runner will cause problems for users.  I've never used spring (and don't have easy access to a project that does) but that's a big unknown given it's use of forking and the potential for them to step on each other's toes.  If the forking runner is incompatible with spring, we probably don't want to default to the fork runner (or find a way to detect that `rspec` was already loaded in a pre-forking runner like `spring`--we could probably stash the original `Process.pid` in a global variable when `lib/rspec/core.rb` first loads, then check that later to see if we're in the same process when we start bisecting).

I'm also not sure how much of a problem the "setup logic in `main` scope of `spec_helper` vs in a `before(:suite)` hook" issues is...as I said, I tried to trigger that condition and failed.  Logically, it's easy to understand how a problem could occur there but actually triggering it on purpose is trickier.

> we could do a deprecation style warning for not setting which one you're using and explain the choice

We could go this route, but I'd really prefer not to. It's kinda the worst option in my mind.  Setting the bisect runner is a little tricky compared to a normal config option.  The standard way people did config before `rspec --init` generated `--require spec_helper` would not even work to allow you to set the bisect runner.  Historically, many projects didn't put a `--require` option in `.rspec`, and instead manually required `spec_helper` from each spec file.  Such an approach would not work to set the bisect runner, as `--bisect` does not load any spec files before instantiating the runner.  The config option only works if it is set in a file loaded via a `--require` option.

After writing that up, I'm realizing it would be beneficial to users if we warned or errored when `config.bisect_runner = runner_option` has no effect due to it not being a `--require` file.  Thoughts?