@JonRowe 2018-01-28
> After writing that up, I'm realizing it would be beneficial to users if we warned or errored when config.bisect_runner = runner_option has no effect due to it not being a --require file. Thoughts?

That seems like a good idea!