@myronmarston 2018-02-09
OK, @JonRowe, I implemented the change we talked about and also added some more output when the `:fork` runner is used on a test suite that has problems with it.  For future reference, here's some example `spec_helper` code that triggers this problem:

``` ruby
RSpec.configure do |config|
  File.write("/tmp/some_resource", "in_use")

  config.after(:suite) do
    File.exist?("/tmp/some_resource") && File.delete("/tmp/some_resource")
  end

  config.before do
    unless File.exist?("/tmp/some_resource")
      raise "file missing"
    end
  end
end
```

The fix for this problem is to use the `:shell` runner (as explained by the text I added) or move the `File.write` into a `before(:suite)` hook so it happens on each subset run (but that's way harder to explain so I didn't put that in the warning).

Thoughts?